Check if a child theme is loaded before enqueueing css. fixes #230
This addresses the issue in Storefront 1.5.1 when using a child theme
and RTL. Prior to this, the rtl stylesheet would not be loaded. It is
now loaded regardless when a child theme is active.

An added benefit of this is that child themes no longer need to import
style.css manually, which is bad for performance anyway.