Update storefront-template-functions.php
We switched our theme to storefront, but I was a bit surprised that this theme didn't support customization more than it is at this moment, for example, we want to override the output from "default" hardcoded functions. there seems to be no filters at all. The only ability was to redefine the functions on the run-time (which is quite inflexible, for example we want to control the filter during execution to conditionally control the output, which is impossible with functions.  Also, it's impossible with functions, that they were overriden from 2 different plugins, because the "function already defined" problems arise.  So, it's better to use wordpress filters).
It would be nice if in addition of this modification, you later implemented filters to control the output of `echo` commands too.