Exclude storefront/ and storefront.zip when preparing a release
This commit makes sure that storefront/ and storefront.zip are
not included in the list of files that are copied to the directory used
to create a production ready version of the theme using the command
`npm run deploy`. This is necessary to avoid issues like the one that
happened in the 2.5.2 version where zip file generated for the release
included a duplicated copy of Storefront inside it.