Remove `.gitattributes` file and add excluded files to `composer.json` (#1763)
* Remove .gitattributes file

This file is no longer needed because it affects users who want to download the archives from GitHub. If we want to exclude files from the built archives then we can use the composer.json file to do this.

* Add files to exclude to composer.json

These source/development config files shouldn't be in the built archive, so we will exclude them from the composer.json file. We need to negate the exclusion because some files are ignored by .gitignore.